feat(notmuch): show originating folder in index
Changes the behavior of '%b' in 'index_format' to display the email's
folder instead of the notmuch query if it exists. In a notmuch mailbox,
the current behavior of index_format's %b' expando displays the query
portion of a notmuch mailbox declaration.

The reference documentation for 'index_format' says '%b' shows "Filename
of the original message folder (think mailbox)". By my interpretation,
displaying the notmuch query is incorrect as it _is not_ the original
message folder. The underlying maildir folder _is_ the original message
folder. Therefore it should be shown instead.

[0] https://neomutt.org/guide/reference.html#3-169-%C2%A0index_format